Backend/requirements.txt
absl-py==2.1.0
asgiref==3.7.2
astroid==3.0.2
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dill==0.3.8
Django==5.0.3
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dm-tree==0.1.8
flatbuffers==24.3.7
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
install==1.3.5
isort==5.13.2
joblib==1.3.2
keras==3.0.5
libclang==18.1.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
nltk==3.8.1
numpy==1.26.4
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.1
pillow==10.2.0
platformdirs==4.2.0
protobuf==4.25.3
Pygments==2.17.2
PyJWT==2.8.0
pylint==3.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2023.12.25
requests==2.31.0
rest-framework-simplejwt==0.0.2
rich==13.7.1
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
threadpoolctl==3.3.0
tomlkit==0.12.3
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2023.4
urllib3==2.2.1
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
